+++ Tests/SwiftLintFrameworkTests/FunctionArgumentsSpacingRuleTests.swift
@@ -1,8 +1,0 @@
-@testable import SwiftLintBuiltInRules
-
-class FunctionArgumentsSpacingRuleTests: SwiftLintTestCase {
-    func testFunctionArgumentsSpacingRule() {
-        let description = FunctionArgumentsSpacingRule.description
-        verifyRule(description)
-    }
-}