+++ Tests/SwiftLintFrameworkTests/NoUnnecessarySpacesRuleTests.swift
@@ -1,0 +1,8 @@
+@testable import SwiftLintBuiltInRules
+
+class NoUnnecessarySpacesRuleTests: SwiftLintTestCase {
+    func testNoUnnecessarySpacesRule() {
+        let description = NoUnnecessarySpacesRule.description
+        verifyRule(description)
+    }
+}